+++ packages/core/src/__test__/AbstractAppliance.test.js
@@ -4,6 +4,7 @@ import {
 	Payload,
 	PayloadArray,
 } from '@tvkitchen/base-classes'
+import { applianceEvents } from '@tvkitchen/base-constants'
 import AbstractAppliance from '../AbstractAppliance'
 import {
 	FullyImplementedAppliance,
@@ -87,6 +88,22 @@ describe('AbstractAppliance #unit', () => {
 			)
 			implementedAppliance.invoke = jest.fn().mockReturnValueOnce(remainingPayloads)
 			expect(await implementedAppliance.ingestPayload(payload)).toBe(false)
+		})
+	})
+
+	describe('emit', () => {
+		it('should emit an event of the specified type', () => {
+			const implementedAppliance = new FullyImplementedAppliance()
+			const emitSpy = jest.spyOn(implementedAppliance.emitter, 'emit')
+			implementedAppliance.emit(applianceEvents.STARTING)
+			expect(emitSpy).toBeCalledWith(applianceEvents.STARTING)
+		})
+		it('should emit an event with specified args', () => {
+			const implementedAppliance = new FullyImplementedAppliance()
+			const payload = new Payload()
+			const emitSpy = jest.spyOn(implementedAppliance.emitter, 'emit')
+			implementedAppliance.emit(applianceEvents.PAYLOAD, payload)
+			expect(emitSpy).toBeCalledWith(applianceEvents.PAYLOAD, payload)
 		})
 	})
 })