+++ packages/core/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial implementation of the AbstractAppliance.
 - Add the `base-constants` dependency for applianceEvents.
+- Add `emit` method to AbstractAppliance.
 
 ### Changed
 - Increase the `base-classes` version dependency to 1.3.0.